+++ openedx_events/event_bus/avro/tests/test_utilities.py
@@ -1,12 +1,10 @@
 """
 Utility methods and classes for testing various modules in event_bus.avro.
 """
-import io
 import re
 from datetime import datetime
 
 import attr
-import fastavro
 from opaque_keys.edx.keys import CourseKey
 
 from openedx_events.event_bus.avro.custom_serializers import BaseCustomTypeAvroSerializer
@@ -27,42 +25,6 @@ def create_simple_signal(data_dict):
         event_type="simple.signal",
         data=data_dict
     )
-
-
-def serialize_event_data_to_bytes(event_data, signal):
-    """
-    Utility method to make sure an Avro serializer can actually serialize given a schema and data
-    to serialize
-
-    Arguments:
-        event_data: Event data to be sent via an OpenEdxPublicSignal's send_event method
-        signal: An instance of OpenEdxPublicSignal
-    Returns:
-        bytes: Byte representation of the event_data, to be sent over the wire
-    """
-    serializer = AvroSignalSerializer(signal)
-    schema_dict = serializer.schema
-    out = io.BytesIO()
-    data_dict = serializer.to_dict(event_data)
-    fastavro.schemaless_writer(out, schema_dict, data_dict)
-    out.seek(0)
-    return out.read()
-
-
-def deserialize_bytes_to_event_data(bytes_from_wire, signal):
-    """
-    Utility method to make sure an Avro deserializer can actually deserialize given a event_bus and Avro-serialized
-    data
-
-    Arguments:
-        bytes_from_wire: data that was serialized by an Avro serializer
-        signal: An instance of OpenEdxPublicSignal
-    """
-    deserializer = AvroSignalDeserializer(signal)
-    schema_dict = deserializer.schema
-    data_file = io.BytesIO(bytes_from_wire)
-    as_dict = fastavro.schemaless_reader(data_file, schema_dict)
-    return deserializer.from_dict(as_dict)
 
 
 # Useful simple attr classes